requirements.txt
PyQt5
pandas
scikit-learn
requests
Pillow
